Revert all changes re: moving content to a Git submodule
Probably not the wisest idea to move content to a subdirectory, since the site have to be rebuilt to update the content anyway. Or that would also require some webhook setup, which, while it can be done via the hut CLI, is a burden to use and not officially supported by the platform.